Move Checkout Header template part tests to parameterized tests

+++ plugins/woocommerce-blocks/tests/e2e/tests/templates/template-customization.block_theme.side_effects.spec.ts
@@ -18,6 +18,9 @@ const userText = 'Hello World in the template';
 const woocommerceTemplateUserText = 'Hello World in the WooCommerce template';
 
 CUSTOMIZABLE_WC_TEMPLATES.forEach( ( testData ) => {
+	if ( ! testData.canBeOverridenByThemes ) {
+		return;
+	}
 	test.describe( `${ testData.templateName } template`, async () => {
 		test.afterAll( async () => {
 			await deleteAllTemplates( 'wp_template' );
